first linux

+++ common/create_py.py
@@ -11,7 +11,6 @@
 import os
 import re
 import glob
-
 from common import setting
 
 
@@ -20,6 +19,7 @@ def create_py(yamlDir="", whichYaml=None, pyPath=""):
     with open(setting.CASE_TEMPLATE, encoding='utf-8') as fr:
         src_content = fr.read()
     dire = yamlDir.split("\\")
+    print(dire)
     if len(yamlDir.split("\\")) > 1:
         yamlDir, secondDir = yamlDir.split("\\")
         yamlPath = os.path.join(setting.DATA_PATH, yamlDir, secondDir)
@@ -27,7 +27,7 @@ def create_py(yamlDir="", whichYaml=None, pyPath=""):
         yamlPath = os.path.join(setting.DATA_PATH, yamlDir)
 
     all_file = glob.glob(yamlPath + os.sep + '*.yaml')
-
+    print(all_file)
     if whichYaml:
         all_file = [all_file[whichYaml-1]]
     else:
@@ -50,17 +50,20 @@ def create_py(yamlDir="", whichYaml=None, pyPath=""):
         # case目录下创建python文件
         py_path = os.path.join(casePath, "test_" + class_name.lower() + '.py')
 
+        if os.path.exists(py_path):
+            pass
+        else:
         # 向case目录下的python文件中写入内容
-        open(py_path, 'w', encoding='utf-8').write(py_content)
+            open(py_path, 'w', encoding='utf-8').write(py_content)
 
-        print("yaml文件>>>>>>>>>>>>>>>>>>>%s" % file)
-        print("生成的py文件>>>>>>>>>>>>>>>>>>>%s" % py_path)
+            print("yaml文件>>>>>>>>>>>>>>>>>>>%s" % file)
+            print("生成的py文件>>>>>>>>>>>>>>>>>>>%s" % py_path)
 
 
 if __name__ == "__main__":
     yamlPath = "登录"  # yamlCase下面模块路径，为空表示选择的yaml文件yamlCase根目录下*.yaml文件
     whichYaml = None  # None yamlPath 目录下所有的yaml文件都生成py,1表示目录下第一个文件
-    pyPath = ""        # case下面模块路径,为空表示生成py文件放在case根目录
+    pyPath = "a登录"        # case下面模块路径,为空表示生成py文件放在case根目录
     create_py(yamlPath, whichYaml, pyPath)
